product_service/requirements.txt
alembic==1.12.1
blinker==1.6.3
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
passlib==1.7.4
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.1
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
Werkzeug==3.0.1
WTForms==3.1.0
